fix: bug using useStorage
